Adding ruleset for Api, Collector.AspNet.Win, Collector.AspNetCore, Collector.AspNetCore.Tests
