+++ test/OpenTelemetry.Collector.AspNetCore.Tests/IncomingRequestsCollectionsIsAccordingToTheSpecTests.cs
@@ -14,18 +14,18 @@
 // limitations under the License.
 // </copyright>
 
-using OpenTelemetry.Trace.Configuration;
-using Xunit;
-using Microsoft.AspNetCore.Mvc.Testing;
-using System.Threading.Tasks;
-using Microsoft.Extensions.DependencyInjection;
-using OpenTelemetry.Trace;
-using OpenTelemetry.Trace.Export;
-using Moq;
-using Microsoft.AspNetCore.TestHost;
 using System;
+using System.Threading.Tasks;
 using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.Mvc.Testing;
+using Microsoft.AspNetCore.TestHost;
+using Microsoft.Extensions.DependencyInjection;
+using Moq;
+using OpenTelemetry.Trace;
+using OpenTelemetry.Trace.Configuration;
+using OpenTelemetry.Trace.Export;
 using TestApp.AspNetCore._3._1;
+using Xunit;
 
 namespace OpenTelemetry.Collector.AspNetCore.Tests
 {
@@ -37,18 +37,6 @@ namespace OpenTelemetry.Collector.AspNetCore.Tests
         public IncomingRequestsCollectionsIsAccordingToTheSpecTests(WebApplicationFactory<Startup> factory)
         {
             this.factory = factory;
-
-        }
-
-        public class TestCallbackMiddlewareImpl : CallbackMiddleware.CallbackMiddlewareImpl
-        {
-
-            public override async Task<bool> ProcessAsync(HttpContext context)
-            {
-                context.Response.StatusCode = 503;
-                await context.Response.WriteAsync("empty");
-                return false;
-            }
         }
 
         [Fact]
@@ -69,7 +57,6 @@ namespace OpenTelemetry.Collector.AspNetCore.Tests
                     }))
                 .CreateClient())
             {
-
                 try
                 {
                     // Act
@@ -88,7 +75,7 @@ namespace OpenTelemetry.Collector.AspNetCore.Tests
                     }
 
                     // We need to let End callback execute as it is executed AFTER response was returned.
-                    // In unit tests environment there may be a lot of parallel unit tests executed, so 
+                    // In unit tests environment there may be a lot of parallel unit tests executed, so
                     // giving some breezing room for the End callback to complete
                     await Task.Delay(TimeSpan.FromSeconds(1));
                 }
